Rename files and make adjustments for sonarcloud
